+++ src/main/java/com/johnymuffin/jvillage/beta/commands/village/JDeleteCommand.java
@@ -14,6 +14,7 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
+import java.math.BigDecimal;
 import java.util.logging.Level;
 
 import static com.johnymuffin.jvillage.beta.JVUtility.cordsInChunk;
@@ -77,7 +78,7 @@ public class JDeleteCommand extends JVBaseCommand implements CommandExecutor {
 
         if (refundAmount > 0 && plugin.isZCoreEnabled()) {
             try {
-                Economy.addBalance(player.getUniqueId(), refundAmount);
+                Economy.addBalance(player.getUniqueId(), BigDecimal.valueOf(refundAmount));
                 this.plugin.logger(Level.INFO, "Successfully refunded $" + refundAmount + " to " + player.getName() + " for deleting village " + village.getTownName());
             } catch (Throwable e) {
                 this.plugin.logger(Level.WARNING, "Failed to refund $" + refundAmount + " to " + player.getName() + " for deleting village " + village.getTownName());
